map and verify age works now
